requirements.txt
fastapi
uvicorn[standard]
python-dotenv     # only if you’re using .env for local dev

